string: refuse a size the length field cannot count
A string object records its length in a field narrower than the size its
constructor takes -- sixteen bits unless the interpreter was built for
large objects. The size was assigned to that field unchecked, while the
allocation and the copy both used the whole of it. So the storage was
there and correct, and the object reported the length the field had
wrapped to: the bytes past the wrap were unreachable, and every reader
answered for the shorter string.

Where that reached furthest was the name of the program to run. A name
of seventy thousand characters became a string of four and a half
thousand, and the shortened name was the one opened. The caller asked
for one file and the interpreter opened another, chosen by whoever
supplied the name -- demonstrated with a name whose leading characters
were a second, existing file, which ran.

Refuse the size in the constructor, so no caller can be given a string
that misreports itself, and state the length at each of the five places
that build one out of an argument the caller supplied -- the program
name, a definition, a resource directory, a device's mode selector and
the output file name. Each of those can then say which of them refused,
which the constructor cannot. Two of them wrote through the result with
no test at all, one of them into the string's own storage.

The resource directory tested its construction against invalidtype,
which a refusal never answers with: a refusal is a null. It now tests
for the type it wants rather than for one of the ways of not having it.

Also widen what the file system's counts travel in. A size and a
position were carried in long, which is 32 bits on LLP64, so both were
spent before the check that weighs them against the integer that must
carry them -- the check was correct and could not fire, because the
number it would refuse had already wrapped on the way in. The stream
methods now count a position in the width the file system does.

The boundary is stated as the field's own edge. It is only reached on
the narrow build: on the large-object build the edge is four gigabytes,
and the test says it leaves it unexercised rather than approximating it
with a shorter name that would prove nothing about the edge.

Source code generated by Claude (Anthropic).
Commit message generated by Claude (Anthropic).

+++ src/lib/xpost_op_file.c
@@ -998,10 +998,10 @@ int xpost_op_file_status (Xpost_Context *ctx,
    three thousand million bytes narrows to a negative length. So it is
    refused rather than reported. */
 static
-int _as_integer(long v, integer *out)
+int _as_integer(long long v, integer *out)
 {
     *out = (integer)v;
-    return (long)*out == v;
+    return (long long)*out == v;
 }
 
 /* string  status  pages bytes referred created true | false
@@ -1016,7 +1016,7 @@ int xpost_op_string_status (Xpost_Context *ctx,
                             Xpost_Object S)
 {
     char *sbuf;
-    long pages, bytes, referred, created;
+    long long pages, bytes, referred, created;
     integer ipages, ibytes, ireferred, icreated;
     int exists;
 
@@ -1259,7 +1259,7 @@ static
 int xpost_op_fileposition (Xpost_Context *ctx,
                            Xpost_Object F)
 {
-    long pos;
+    long long pos;
     integer ipos;
     if (!xpost_file_get_status(ctx->lo, F))
         return ioerror;
